@@ -5,33 +5,36 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.7.18</version>
+    <version>3.2.3</version>
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
   <groupId>uk.gov.dwp.health</groupId>
   <artifactId>ms-account-manager</artifactId>
-  <version>5.5.0</version>
+  <version>5.7.0</version>
   <name>ms-account-manager</name>
   <description>Micro service provides account management &amp; authorization</description>
 
   <properties>
-    <java.version>11</java.version>
+    <java.version>17</java.version>
+    <required-java.version>17</required-java.version>
+    <required-maven.version>3.6.0</required-maven.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <client.base.package.name>uk.gov.dwp.health.account.manager</client.base.package.name>
-    <dwp.message-broker.version>2.12.0-cce5f733</dwp.message-broker.version>
-    <dwp-mongo-change-stream-starter.version>3.2.0-dd59aa6b</dwp-mongo-change-stream-starter.version>
+    <dwp.message-broker.version>3.4.0-bd00860c</dwp.message-broker.version>
+    <dwp-mongo-change-stream-starter.version>4.2.0-359f3f54</dwp-mongo-change-stream-starter.version>
     <dwp.postcode-validation.version>1.3.0-f0fc5e75</dwp.postcode-validation.version>
-    <dwp-meta-data-logger.version>2.0.1-b4ab9a22</dwp-meta-data-logger.version>
-    <swagger-annotations.version>1.6.12</swagger-annotations.version>
+    <dwp-meta-data-logger.version>2.0.0-f17eb705</dwp-meta-data-logger.version>
+    <swagger-annotations.version>2.2.20</swagger-annotations.version>
     <openapi-codegen-maven-plugin.version>6.6.0</openapi-codegen-maven-plugin.version>
-    <mongock.version>4.1.19</mongock.version>
+    <mongock.version>5.3.4</mongock.version>
     <jackson-nullable.version>0.2.6</jackson-nullable.version>
     <lidalia.version>1.2.0</lidalia.version>
-    <maven.compiler.version>3.10.1</maven.compiler.version>
-    <plugin.jacoco.version>0.8.8</plugin.jacoco.version>
+    <maven.compiler.version>3.11.0</maven.compiler.version>
+    <plugin.jacoco.version>0.8.10</plugin.jacoco.version>
     <plugin.owasp.version>7.4.4</plugin.owasp.version>
-    <plugin.spotbugs.version>4.0.0</plugin.spotbugs.version>
+    <plugin.spotbugs.version>4.7.3.5</plugin.spotbugs.version>
+    <spotbugs.version>4.7.3</spotbugs.version>
     <plugin.projectinfo.version>3.4.2</plugin.projectinfo.version>
     <plugin.surefire.version>2.22.2</plugin.surefire.version>
     <plugin.pmd.version>3.20.0</plugin.pmd.version>
