@@ -12,8 +12,16 @@ public class UrlBuilderUtil {
     return baseURI + "/v5/accounts";
   }
 
+  public static String postAccountV7Url() {
+    return baseURI + "/v7/accounts";
+  }
+
   public static String patchClaimantDetailsUrl() {
     return baseURI + "/v4/account/claimantdetails/";
+  }
+
+  public static String patchClaimantDetailsV7Url() {
+    return baseURI + "/v7/account/claimantdetails/";
   }
 
   public static String patchClaimantTransferStatusUrl() {
@@ -24,20 +32,28 @@ public class UrlBuilderUtil {
     return baseURI + "/v4/account/details/id/" + accountRef + "";
   }
 
+  public static String getAccountByIDV7Url(String accountRef) {
+    return baseURI + "/v7/account/details/id/" + accountRef + "";
+  }
+
   public static String getAccountByEmailUrl(String email) {
     return baseURI + "/v4/account/details/email/" + email + "";
+  }
+
+  public static String getAccountByEmailV7Url(String email) {
+    return baseURI + "/v7/account/details/email/" + email + "";
   }
 
   public static String getAccountByNinoUrl(String nino) {
     return baseURI + "/v4/account/details/nino/" + nino + "";
   }
 
-  public static String getMobileNumbersByClaimantIdUrl(String claimantIds) {
-    return baseURI + "/v3/account/mobilephone/claimantid/" + claimantIds;
+  public static String getAccountByNinoV7Url(String nino) {
+    return baseURI + "/v7/account/details/nino/" + nino + "";
   }
 
-  public static String getLimiterUrl() {
-    return baseURI + "/v5/limiter";
+  public static String getMobileNumbersByClaimantIdUrl(String claimantIds) {
+    return baseURI + "/v3/account/mobilephone/claimantid/" + claimantIds;
   }
 
   public static String postIdentifyUrl() {
