@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.dwp.health.account.manager.http.CheckCanApplyHttpRespStatusHandler;
-import uk.gov.dwp.health.logging.OutgoingInterceptor;
+import uk.gov.dwp.health.monitoring.interceptor.OutgoingInterceptor;
 
 import java.util.List;
 
