@@ -1,6 +1,6 @@
 package uk.gov.dwp.health.account.manager.config;
 
-import com.github.cloudyrock.spring.v5.EnableMongock;
+import io.mongock.runner.springboot.EnableMongock;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
