@@ -1,6 +1,6 @@
 package uk.gov.dwp.health.account.manager.migration;
 
-import com.github.cloudyrock.mongock.driver.mongodb.springdata.v3.decorator.impl.MongockTemplate;
+import com.github.cloudyrock.mongock.driver.mongodb.springdata.v4.decorator.impl.MongockTemplate;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
